Add async I/O support
This commit updates the minimum required Zig version to O.16.0-dev
and introduces support for async I/O.

The client interface has changed to accept only a reader and a writer,
instead of a connection. This was done to minimize the amount of things
the client has to worry about, leaving to the user the duty to decide how
to handle reconnection. This might change in the future but for now it's
better to give the user full control rather than have a half-assed
implementation that makes it harder for the user to do the right thing.

The other important change is that now, by integrating with Zig's new Io
interface, the client is fully threadsafe (it was not before, if that's
news to you, go fix your code :^) and implements cross-thread pipelining,
something that I suspect not many Redis clients implement, if any.

There might also bugs included with these changes, please do try hard to
provide a repro if you encounter any. Especially if you experience
concurrency issues that lead to deadlocks.